backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.7
psycopg2-binary==2.9.5  # PostgreSQL adapter
sqlalchemy-utils==0.41.1  # Added for database existence checks
pydantic==2.11.7  # Updated to V2 for modern features
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0  # For environment variable management
sentence-transformers==3.3.1  # Updated for better compatibility
tf-keras==2.19.0  # Required for Keras 3 compatibility with transformers
numpy==1.24.2
pyjwt==2.6.0
bcrypt==4.0.1
alembic==1.10.3
email-validator==2.0.0
requests==2.28.2
beautifulsoup4==4.12.2
tenacity==8.2.2  # Added for retry logic
aiohttp==3.8.4  # For async HTTP requests in enhanced search

# New dependencies for AI-powered features
openai==1.3.8  # For OpenAI API integration (optional)
google-generativeai==0.3.2  # For Gemini API integration (optional)
scikit-learn==1.3.2  # For collaborative filtering and clustering
pandas==2.0.3  # For data analysis and user behavior tracking
matplotlib==3.7.1  # For generating charts and visualizations (optional)
seaborn==0.12.2  # For enhanced visualizations (optional)
